split out forge config
